- event.h: replaced shorts with ints
git-svn-id: svn://svn.ircd-hybrid.org/svnroot/ircd-hybrid/trunk@4336 82007160-df01-0410-b94d-b575c5fd34c7

+++ include/event.h
@@ -40,12 +40,12 @@ struct event
   const char *name;
   event_handler handler;
   time_t when;
-  unsigned short oneshot;
+  unsigned int oneshot;
 
   /* private */
   time_t next;
   void *data;
-  unsigned short enabled;
+  unsigned int enabled;
   dlink_node node;
 };
